Extract calculators as independent from fe code
