+++ src/calculators/prerequisiteOperations.ts
@@ -1,24 +1,24 @@
-import { isPrerequisite, Prerequisite } from '../../definitions/Prerequisites'
-import { Duration } from '../../definitions/auxiliary/Duration'
-import { Combination, isCombination } from '../../definitions/auxiliary/Combination'
+import { isPrerequisite, Prerequisite } from '../definitions/Prerequisites'
+import { Duration } from '../definitions/auxiliary/Duration'
+import { Combination, isCombination } from '../definitions/auxiliary/Combination'
 import {
     languageTestItemValues,
     LanguageTestResult,
     LanguageTestScoreSet
-} from '../../definitions/auxiliary/LanguageTest'
+} from '../definitions/auxiliary/LanguageTest'
 import {
     LanguagePrereq,
     LanguagePrereqResult,
     LanguagePrereqScoreSet
-} from '../../definitions/Prerequisites/LanguagePrereq'
-import { WorkExperienceQuality } from '../../definitions/Qualities/WorkExperience'
-import { WorkExperiencePrereq } from '../../definitions/Prerequisites/WorkExperiencePrereq'
-import { RegionId } from '../../definitions/auxiliary/Region'
-import { ArithmeticComparisonOperator, Interval } from '../../definitions/auxiliary/Operator'
-import { EducationQuality, getEducationStageRank } from '../../definitions/Qualities/EducationExperience'
-import { EducationPrereq } from '../../definitions/Prerequisites/EducationPrereq'
-import { Person } from '../../definitions/Person'
-import { clone } from './clone'
+} from '../definitions/Prerequisites/LanguagePrereq'
+import { WorkExperienceQuality } from '../definitions/Qualities/WorkExperience'
+import { WorkExperiencePrereq } from '../definitions/Prerequisites/WorkExperiencePrereq'
+import { RegionId } from '../definitions/auxiliary/Region'
+import { ArithmeticComparisonOperator, Interval } from '../definitions/auxiliary/Operator'
+import { EducationQuality, getEducationStageRank } from '../definitions/Qualities/EducationExperience'
+import { EducationPrereq } from '../definitions/Prerequisites/EducationPrereq'
+import { Person } from '../definitions/Person'
+import { clone } from '../client/utils/clone'
 
 const DEFAULT_RESULT = true
